fix or silence compiler warnings about narrowing conversions or unused formal parameters (msvc)
(cherry picked from commit 4d2c7079dc09e5f0702496c08f67cbf8e7dc71af)

+++ test/is_permutation_test1.cpp
@@ -113,7 +113,7 @@ void test_sequence1 () {
     
     v.clear ();
     for ( std::size_t i = 5; i < 15; ++i )
-        v.push_back ( i );
+        v.push_back ( static_cast<int>(i) );
     v1 = v;
     BOOST_CHECK ( ba::is_permutation ( v.begin (), v.end (), v.begin ()));  // better be a permutation of itself!
     BOOST_CHECK ( ba::is_permutation ( v.begin (), v.end (), v1.begin ()));
